Resolve merge conflict
